+++ lib/protocol/grpc/call.rb
@@ -36,6 +36,15 @@ module Protocol
 				@metadata ||= Methods.extract_metadata(@request.headers)
 			end
 			
+			# Get the timeout requested by the client.
+			# @returns [Numeric | Nil] The original timeout in seconds, or `nil` if no timeout was specified.
+			def timeout
+				value = @request.headers["grpc-timeout"]
+				value = value.first if value.is_a?(Array)
+				
+				Methods.parse_timeout(value)
+			end
+			
 			# Check if the deadline has expired.
 			# @returns [Boolean] `true` if the deadline has expired, `false` otherwise
 			def deadline_exceeded?